+++ src-ui/app/components/event-log-entry/entry-parsers/changed-vrchat-mic-mute-state.ts
@@ -1,11 +1,7 @@
 import { EventLogEntryParser } from '../event-log-entry-parser';
-import {
-  EventLogBase,
-  EventLogChangedVRChatMicMuteState,
-  EventLogType,
-} from '../../../models/event-log-entry';
+import { EventLogChangedVRChatMicMuteState, EventLogType } from '../../../models/event-log-entry';
 
-export class EventLogChangedVRChatMicMuteStateEntryParser extends EventLogEntryParser<EventLogBase> {
+export class EventLogChangedVRChatMicMuteStateEntryParser extends EventLogEntryParser<EventLogChangedVRChatMicMuteState> {
   entryType(): EventLogType {
     return 'changedVRChatMicMuteState';
   }